# precomp: use normalized extended points