--- a/banderwagon/precomp.go
+++ b/banderwagon/precomp.go
@@ -61,20 +61,24 @@ func NewPrecompMSM(points []Element) (MSMPrecomp, error) {
 // MSM calculates the 256-MSM of the given scalars on the fixed basis.
 // It automatically detects how many non-zero scalars there are and parallelizes the computation.
 func (msm *MSMPrecomp) MSM(scalars []fr.Element) Element {
-	result := Identity.inner
+	result := bandersnatch.IdentityExt
 
 	for i := range scalars {
 		if !scalars[i].IsZero() {
 			msm.precompPoints[i].ScalarMul(scalars[i], &result)
 		}
 	}
-	return Element{inner: result}
+	return Element{inner: bandersnatch.PointProj{
+		X: result.X,
+		Y: result.Y,
+		Z: result.Z,
+	}}
 }
 
 // PrecompPoint is a precomputed table for a single point.
 type PrecompPoint struct {
 	windowSize int
-	windows    [][]bandersnatch.PointAffine
+	windows    [][]bandersnatch.PointExtendedNormalized
 }
 
 // NewPrecompPoint creates a new PrecompPoint for the given point and window size.
@@ -88,27 +92,23 @@ func NewPrecompPoint(point Element, windowSize int) (PrecompPoint, error) {
 
 	res := PrecompPoint{
 		windowSize: windowSize,
-		windows:    make([][]bandersnatch.PointAffine, 256/windowSize),
+		windows:    make([][]bandersnatch.PointExtendedNormalized, 256/windowSize),
 	}
 
-	windows := make([][]bandersnatch.PointProj, 256/windowSize)
+	windows := make([][]bandersnatch.PointExtended, 256/windowSize)
 	group, _ := errgroup.WithContext(context.Background())
 	group.SetLimit(runtime.NumCPU())
 	for i := 0; i < len(res.windows); i++ {
 		i := i
-		base := point.inner
+		base := bandersnatch.PointExtendedFromProj(&point.inner)
 		group.Go(func() error {
-			windows[i] = make([]bandersnatch.PointProj, 1<<(windowSize-1))
+			windows[i] = make([]bandersnatch.PointExtended, 1<<(windowSize-1))
 			curr := base
 			for j := 0; j < len(windows[i]); j++ {
 				windows[i][j] = curr
 				curr.Add(&curr, &base)
 			}
-			batchProjToAffine(windows[i])
-			res.windows[i] = make([]bandersnatch.PointAffine, 1<<(windowSize-1))
-			for j := range windows[i] {
-				res.windows[i][j].FromProj(&windows[i][j])
-			}
+			res.windows[i] = batchNormalizeExtendedPoint(windows[i])
 			return nil
 		})
 		point.ScalarMul(&point, &specialWindow)
@@ -121,12 +121,12 @@ func NewPrecompPoint(point Element, windowSize int) (PrecompPoint, error) {
 // ScalarMul multiplies the point by the given scalar using the precomputed points.
 // It applies a trick to push a carry between windows since our precomputed tables
 // avoid storing point inverses.
-func (pp *PrecompPoint) ScalarMul(scalar fr.Element, res *bandersnatch.PointProj) {
+func (pp *PrecompPoint) ScalarMul(scalar fr.Element, res *bandersnatch.PointExtended) {
 	numWindowsInLimb := 64 / pp.windowSize
 
 	scalar.FromMont()
 	var carry uint64
-	var pNeg bandersnatch.PointAffine
+	var pNeg bandersnatch.PointExtendedNormalized
 	for l := 0; l < fr.Limbs; l++ {
 		for w := 0; w < numWindowsInLimb; w++ {
 			windowValue := (scalar[l]>>(pp.windowSize*w))&((1<<pp.windowSize)-1) + carry
@@ -139,11 +139,11 @@ func (pp *PrecompPoint) ScalarMul(scalar fr.Element, res *bandersnatch.PointProj
 				windowValue = (1 << pp.windowSize) - windowValue
 				if windowValue != 0 {
 					pNeg.Neg(&pp.windows[l*numWindowsInLimb+w][windowValue-1])
-					res.MixedAdd(res, &pNeg)
+					bandersnatch.ExtendedAddNormalized(res, res, &pNeg)
 				}
 				carry = 1
 			} else {
-				res.MixedAdd(res, &pp.windows[l*numWindowsInLimb+w][windowValue-1])
+				bandersnatch.ExtendedAddNormalized(res, res, &pp.windows[l*numWindowsInLimb+w][windowValue-1])
 			}
 		}
 	}
@@ -190,6 +190,52 @@ func batchProjToAffine(points []bandersnatch.PointProj) []bandersnatch.PointAffi
 			a := result[i].X
 			result[i].X.Mul(&points[i].X, &a)
 			result[i].Y.Mul(&points[i].Y, &a)
+		}
+	})
+
+	return result
+}
+
+func batchNormalizeExtendedPoint(points []bandersnatch.PointExtended) []bandersnatch.PointExtendedNormalized {
+	result := make([]bandersnatch.PointExtendedNormalized, len(points))
+	zeroes := make([]bool, len(points))
+	accumulator := fp.One()
+
+	// batch invert all points[].Z coordinates with Montgomery batch inversion trick
+	// (stores points[].Z^-1 in result[i].X to avoid allocating a slice of fr.Elements)
+	for i := 0; i < len(points); i++ {
+		if points[i].Z.IsZero() {
+			zeroes[i] = true
+			continue
+		}
+		result[i].X = accumulator
+		accumulator.Mul(&accumulator, &points[i].Z)
+	}
+
+	var accInverse fp.Element
+	accInverse.Inverse(&accumulator)
+
+	for i := len(points) - 1; i >= 0; i-- {
+		if zeroes[i] {
+			// do nothing, (X=0, Y=0) is infinity point in affine
+			continue
+		}
+		result[i].X.Mul(&result[i].X, &accInverse)
+		accInverse.Mul(&accInverse, &points[i].Z)
+	}
+
+	// batch convert to affine.
+	parallel.Execute(len(points), func(start, end int) {
+		for i := start; i < end; i++ {
+			if zeroes[i] {
+				// do nothing, (X=0, Y=0) is infinity point in affine
+				continue
+			}
+
+			a := result[i].X
+			result[i].X.Mul(&points[i].X, &a)
+			result[i].Y.Mul(&points[i].Y, &a)
+			result[i].T.Mul(&result[i].X, &result[i].Y)
 		}
 	})